Review feedback (move to concern + better error messages)
